requirements.txt
Django~=3.2.10
crispy-bootstrap5
django-crispy-forms
django-filter
django-debug-toolbar
django_utils_six
django-mptt
django-extra-views